Load styles into previews with __editorAssets global to prevent issues on Firefox
